Fix crash when switching monitors
When switching monitors AND displaying notifications at the
same time, a rare situation might occur when notification
window will be added to two or more stacks. One of the
stacks will be removed soon enough, but the "destroy"
handler will persist and will be triggered when the window
disappears. This results in a reference to a freed memory
and (usually) a daemon crash, a typical one is
https://retrace.fedoraproject.org/faf/reports/359836/ for
example.

So, the removal handlers referring to a deleted stack must
be cancelled prior to removal.

Closes https://github.com/mate-desktop/mate-notification-daemon/pull/34