coco: fix window height in gtk3 builds
Reduce the height of the notification window in gtk3 builds to one proportionate to the window contents. In most cases this will prevent the notification window from overlapping bottom panels. Not all notifications send the window to the exact same place so a few cases such as notify-send can still overlap a bottom panel but with no loss of usability.