daemon: Export the "action-icons" capability to clients
The freedesktop specification states that clients may opt to use named icons
instead of textual labels on action buttons. In this instance, the icon name
is passed as the 'key' (actions[i+1]) element in the actions data passed
to the daemon.

For themes to render the icon, they should check the hints passed by the
client for a boolean value of "action-icons", and if it is set they should
then create an image button based on the key, and not the label.

Many applications now make use of action-icons, such as media players, to
provide self describing actions with well known icon names, such as playback
controls. In the absence of support for action-icons they fallback to less
useful buttons and generally less actions.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>