themes: Add functionality to test for presence of action-icons
Each of the themes will have to detect the 'action-icons' hint passed to
them in 'set_notification_hints' to make a determination on whether or not
to use image based buttons, instead of textual buttons, in their actions
boxes.

This factory behaviour simply extends the existing checks for urgency-hint,
as well as unifying the code style between the various in-tree theme
implementations.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>