themes: Ensure that we never add more than one pie_countdown
Prior to this commit we would always end up adding multiple pie_countdown
widgets in a number of situations when updating the notification. In modern
media player apps, many updates are sent for the lifetime of a single
notification window, leading to new drawing areas being added and counting
down with each track/information change.

Note this hasn't been noticable for quite some time, because in many of the
themes the pie countdown hasn't actually been visible due to the actions
box having become hidden completely, meaning that the actions and pie show
very briefly, and are immediately nuked on a change to the notification.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>