themes: Implement support for action-icons buttons
Now that all of the elements are in place to support (dynamically) the
action icons, we can use the 'key' in place of the 'label' when the client
sends an hint requesting it.

Some other minor issues were corrected in this change, such as ensuring that
we always show *all* widgets inside of the notification window so that when
a live notification window is updated, the action box state is not then
hidden from view.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>