Fix decoding the hints dictionary.
g_variant_lookup works like g_variant_get, so using a format string of "v" only
works if the type is actually a GVariant. Since none of the hints have GVariant
values, that means every g_variant_get with a "v" format will fail.

Fix all of the g_variant_lookup calls so that they either unpack a value
directly, or use a "@" prefix when it's more convenient to fetch the value as
a GVariant pointer.

In addition, in cases where we do fetch a GVariant, make sure that we
unreference it afterward.