Auto-start m-n-d with session
Relying on D-Bus activation to launch org.freedesktop.Notifications can
result in selecting the wrong implementation if multiple daemons are
installed.

Fix this by launching m-n-d with the session rather than depending on
D-Bus activation.  We keep D-Bus activation for the moment for backward
compatibility and to re-start the daemon in otherwise non-problematic
situations if it crashes.

Fixes #133 and #174.